nonebot_plugin_penguin/__init__.py
from nonebot import require

from . import db, trim, user, types, utils, render, request, startup, schedule

require("nonebot_plugin_apscheduler")
require("nonebot_plugin_saa")


__all__ = [
    "render",
    "db",
    "request",
    "schedule",
    "startup",
    "trim",
    "user",
    "types",
    "utils",
]
