pyproject.toml
[tool.poetry]
name = "nonebot-plugin-penguin"
version = "0.0.1"
description = "get penguin data and send from https://penguin-stats.io/"
authors = ["Azide <rukuy@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "nonebot_plugin_penguin" }]
homepage = "https://github.com/AzideCupric/nonebot-plugin-penguin"
repository = "https://github.com/AzideCupric/nonebot-plugin-penguin"
keywords = []
classifiers = [
      "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "License :: OSI Approved :: MIT License",
]

[tool.poetry.dependencies]
python = "^3.10"
tinydb = "^4.7.1"
nonebot2 = { extras = ["fastapi"], version = "^2.0.0rc4" }
httpx = "^0.23.3"
nonebot-plugin-htmlrender = "^0.2.0.3"
nonebot-plugin-apscheduler = "^0.2.0"
nonebot-plugin-send-anything-anywhere = "^0.2.2"
nonebot-adapter-onebot = "^2.2.2"
freezegun = "^1.2.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^23.1.0"
nonemoji = "^0.1.2"
pre-commit = "^3.1.0"
pytest = "^7.3.0"
nonebug = {git = "https://github.com/nonebot/nonebug.git"}
pytest-mock = "^3.10.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
respx = "^0.20.1"
fastapi = "^0.95.0"
flaky = "^3.7.0"
pytest-xdist = {extras = ["psutil"], version = "^3.2.1"}

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pycln]
path = "."

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"